#642 relocate app folder to be directly under frontend
