+++ src/frontend/app/app-authenticated/app-authenticated.tsx
@@ -4,15 +4,15 @@
  */
 
 import { Route, Switch } from 'react-router-dom';
-import { routes } from '../../../../shared/routes';
-import ChangeRequests from '../../../pages/ChangeRequestsPage/change-requests';
-import Projects from '../../../pages/ProjectsPage/projects';
-import { PageNotFound } from '../../../pages/PageNotFoundPage/page-not-found';
-import Home from '../../../pages/HomePage/home';
-import NavTopBar from '../../nav-top-bar/nav-top-bar';
-import Settings from '../../../pages/SettingsPage/settings';
+import { routes } from '../../../shared/routes';
+import ChangeRequests from '../../pages/ChangeRequestsPage/change-requests';
+import Projects from '../../pages/ProjectsPage/projects';
+import { PageNotFound } from '../../pages/PageNotFoundPage/page-not-found';
+import Home from '../../pages/HomePage/home';
+import NavTopBar from '../../components/nav-top-bar/nav-top-bar';
+import Settings from '../../pages/SettingsPage/settings';
 import styles from './app-authenticated.module.css';
-import Sidebar from '../../sidebar/sidebar';
+import Sidebar from '../../components/sidebar/sidebar';
 
 const AppAuthenticated: React.FC = () => {
   return (
