+++ src/frontend/app/app-context-auth/app-context-auth.tsx
@@ -4,8 +4,8 @@
  */
 
 import { createContext } from 'react';
-import { useProvideAuth } from '../../../../services/auth.hooks';
-import { Auth } from '../../../../shared/types';
+import { useProvideAuth } from '../../../services/auth.hooks';
+import { Auth } from '../../../shared/types';
 import './app-context-auth.module.css';
 
 export const AuthContext = createContext<Auth | undefined>(undefined);
