+++ src/frontend/app/app-context-query/app-context-query.test.tsx
@@ -4,7 +4,7 @@
  */
 
 import { render, screen } from '@testing-library/react'; // avoid circular dependency
-import { useAllChangeRequests } from '../../../../services/change-requests.hooks';
+import { useAllChangeRequests } from '../../../services/change-requests.hooks';
 import AppContextQuery from './app-context-query';
 
 describe('app context', () => {
