+++ src/frontend/app/app-main/app-main.test.tsx
@@ -3,7 +3,7 @@
  * See the LICENSE file in the repository root folder for details.
  */
 
-import { render, screen } from '../../../../test-support/test-utils';
+import { render, screen } from '../../../test-support/test-utils';
 import AppMain from './app-main';
 
 jest.mock('../app-public/app-public', () => {
