+++ src/services/tests/projects.hooks.test.tsx
@@ -6,7 +6,7 @@
 import { renderHook } from '@testing-library/react-hooks';
 import { AxiosResponse } from 'axios';
 import { Project } from 'utils';
-import wrapper from '../../frontend/components/app/app-context-query/app-context-query';
+import wrapper from '../../frontend/app/app-context-query/app-context-query';
 import { mockPromiseAxiosResponse } from '../../test-support/test-data/test-utils.stub';
 import { exampleAllProjects, exampleProject1 } from '../../test-support/test-data/projects.stub';
 import { exampleWbsProject1 } from '../../test-support/test-data/wbs-numbers.stub';
