+++ src/services/tests/users.hooks.test.tsx
@@ -7,7 +7,7 @@ import { renderHook } from '@testing-library/react-hooks';
 import { act } from 'react-dom/test-utils';
 import { AxiosResponse } from 'axios';
 import { User } from '@prisma/client';
-import wrapper from '../../frontend/components/app/app-context-query/app-context-query';
+import wrapper from '../../frontend/app/app-context-query/app-context-query';
 import { mockPromiseAxiosResponse } from '../../test-support/test-data/test-utils.stub';
 import { exampleAllUsers, exampleAdminUser } from '../../test-support/test-data/users.stub';
 import { getAllUsers, getSingleUser, logUserIn } from '../users.api';
